Moved some commands from Misc.py to Server.py (will be used for server related commands like serverinfo and membercount)
